Final cleanup and optimizations
